HB, URR, KTV working

+++ db/migrate/20141230053701_create_analisis_mensuals.rb
@@ -3,14 +3,14 @@ class CreateAnalisisMensuals < ActiveRecord::Migration
     create_table :analisis_mensuals do |t|
       t.decimal :hto
       t.decimal :hb
-      t.integer :peso_pre
-      t.integer :peso_post
-      t.integer :urea_pre
-      t.integer :urea_post
-      t.integer :urr
-      t.integer :ktv
-      t.integer :tgp
-      t.integer :tgo
+      t.decimal :peso_pre
+      t.decimal :peso_post
+      t.decimal :urea_pre
+      t.decimal :urea_post
+      t.decimal :urr
+      t.decimal :ktv
+      t.decimal :tgp
+      t.decimal :tgo
       t.date :fecha
       t.references :paciente
